+++ packages/dataviews/src/dataform-controls/password.tsx
@@ -1,21 +1,15 @@
 /**
  * WordPress dependencies
  */
-import {
-	Button,
-	privateApis,
-	__experimentalInputControlSuffixWrapper as InputControlSuffixWrapper,
-} from '@wordpress/components';
+import { Button } from '@wordpress/components';
 import { useCallback, useState } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
 import { seen, unseen } from '@wordpress/icons';
-
 /**
  * Internal dependencies
  */
+import ValidatedText from './utils/validated-input';
 import type { DataFormControlProps } from '../types';
-import { unlock } from '../lock-unlock';
-
-const { ValidatedInputControl } = unlock( privateApis );
 
 export default function Password< Item >( {
 	data,
@@ -23,72 +17,34 @@ export default function Password< Item >( {
 	onChange,
 	hideLabelFromVision,
 }: DataFormControlProps< Item > ) {
-	const { id, label, placeholder, description } = field;
-	const value = field.getValue( { item: data } );
 	const [ isVisible, setIsVisible ] = useState( false );
-	const [ customValidity, setCustomValidity ] =
-		useState<
-			React.ComponentProps<
-				typeof ValidatedInputControl
-			>[ 'customValidity' ]
-		>( undefined );
-
-	const onChangeControl = useCallback(
-		( newValue: string ) =>
-			onChange( {
-				[ id ]: newValue,
-			} ),
-		[ id, onChange ]
-	);
 
 	const toggleVisibility = useCallback( () => {
 		setIsVisible( ( prev ) => ! prev );
 	}, [] );
 
 	return (
-		<ValidatedInputControl
-			required={ !! field.isValid?.required }
-			onValidate={ ( newValue: any ) => {
-				const message = field.isValid?.custom?.(
-					{
-						...data,
-						[ id ]: newValue,
-					},
-					field
-				);
-
-				if ( message ) {
-					setCustomValidity( {
-						type: 'invalid',
-						message,
-					} );
-					return;
-				}
-
-				setCustomValidity( undefined );
-			} }
-			customValidity={ customValidity }
-			label={ label }
-			placeholder={ placeholder }
-			value={ value ?? '' }
-			help={ description }
-			onChange={ onChangeControl }
-			hideLabelFromVision={ hideLabelFromVision }
-			type={ isVisible ? 'text' : 'password' }
-			suffix={
-				<InputControlSuffixWrapper variant="control">
+		<ValidatedText
+			{ ...{
+				data,
+				field,
+				onChange,
+				hideLabelFromVision,
+				type: isVisible ? 'text' : 'password',
+				suffix: (
 					<Button
 						icon={ isVisible ? unseen : seen }
 						onClick={ toggleVisibility }
 						size="small"
 						variant="tertiary"
 						aria-label={
-							isVisible ? 'Hide password' : 'Show password'
+							isVisible
+								? __( 'Hide password' )
+								: __( 'Show password' )
 						}
 					/>
-				</InputControlSuffixWrapper>
-			}
-			__next40pxDefaultSize
+				),
+			} }
 		/>
 	);
 }